Move link-to tests
